requirements.txt
beautifulsoup4>=4.12.3
requests>=2.31.0
requests-oauthlib>=1.3.1
requests-toolbelt>=1.0.0

pymongo>=4.3.3
nltk>=3.8.1
pyodbc>=5.2.0

numpy>=1.26.4,<2.0
pandas>=2.1.0

faiss-cpu>=1.9.0

langchain>=0.3.24
langchain-community>=0.3.1
langchain-huggingface>=0.1.2

google-generativeai>=0.8.5
gradio>=4.44.1
sentence-transformers>=4.1.0

pypdf>=5.8.0
PyPDF2>=3.0.1

pinecone>=6.0.2
pinecone-client>=6.0.0

fastapi>=0.110
python-dotenv>=1.0.1